% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefglm.R
\name{coef.glm}
\alias{coef.glm}
\title{Extract model coefficients}
\usage{
\method{coef}{glm}(object, type = c("raw", "bias", "corrected"), ...)
}
\arguments{
\item{object}{The fitted model.}

\item{type}{Must be one of \code{raw}, \code{bias}, or \code{corrected}.}

\item{...}{Additional arguments.}
}
\description{
Extract model coefficients.
}
