% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicated.matrix.R
\name{duplicated.matrix}
\alias{duplicated.matrix}
\title{Determine duplicate elements}
\usage{
duplicated.matrix(x, incomparables = FALSE, MARGIN = 1L, fromLast = FALSE,
  ...)
}
\arguments{
\item{x}{A matrix.}

\item{incomparables}{The values that cannot be compared. FALSE means all values can be compared.}

\item{MARGIN}{The array margin to hold fixed.}

\item{fromLast}{Logical. Should duplication be considered from teh last instance?}

\item{...}{Additional arguments.}
}
\description{
Determine which elements are duplicates.
}
