% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicated.matrix.R
\name{unique.matrix}
\alias{unique.matrix}
\title{Remove duplicate elements}
\usage{
unique.matrix(x, incomparables = FALSE, MARGIN = 1, fromLast = FALSE, ...)
}
\arguments{
\item{x}{A matrix.}

\item{incomparables}{The values that cannot be compared. FALSE means all values can be compared.}

\item{MARGIN}{The array margin to hold fixed.}

\item{fromLast}{Logical. Should duplication be considered from teh last instance?}

\item{...}{Additional arguments.}
}
\description{
Remove duplicate elements.
}
