% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinQuasi.R
\name{BinQuasi}
\alias{BinQuasi}
\title{Analyzing Replicated ChIP Sequencing Data Using Quasi-Likelihood}
\description{
Identify peaks in ChIP-seq data with biological replicates.
}
\details{
Identify peaks in ChIP-seq data with biological replicates using a one-sided quasi-likelihood ratio test in quasi-Poisson or quasi-negative binomial models.
}
